ci-tests: adding test output on commandline while xmlreporter is collecting data
git-svn-id: http://svn.akelos.org@971 a2fa5c27-f921-0410-a72c-bf682d381be0

+++ branches/arnoschn/cache/CiReportTask.php
@@ -47,7 +47,7 @@ class CiReportTask extends PHPUnitReportTask
                 $tests += (int)$attributes->tests;
                 $failures += (int)$attributes->failures;
                 $errors += (int)$attributes->errors;
-                $time += (int)$attributes->time;
+                $time += (float)$attributes->time;
             }
             $environment = array();
             $environment['php']=$php;
